+++ src/main/java/christmas/validator/DomainValidator.java
@@ -2,8 +2,8 @@ package christmas.validator;
 
 import static christmas.util.constant.ValidatorConst.COMPARE_EQUAL_VALUE_RESULT;
 import static christmas.util.constant.ValidatorConst.SINGLE_ORDER_CATEGORY;
-import static christmas.util.exception.ErrorMessage.ORDER_ERROR_MESSAGE;
-import static christmas.util.exception.ErrorMessage.PREFIX;
+import static christmas.validator.ErrorMessage.ORDER_ERROR_MESSAGE;
+import static christmas.validator.ErrorMessage.PREFIX;
 
 import christmas.util.Menu;
 import java.math.BigInteger;
